% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents}
\alias{accidents}
\title{Mexico accident dataset}
\format{
A sf object with 2465 rows and 29 variables:
\describe{
  \item{CVEGEO}{the ids.}
  \item{CVE_ENT}{is the numeric key of state.}
  \item{CVE_MUN}{is the numeric key to the Municipality}
  \item{NOMGEO}{is the numeric key to the Sectors}
  \item{ACC_2003}{accident cases from 2003}
  \item{ACC_2004}{accident cases from 2004}
  \item{ACC_2005}{accident cases from 2005}
  \item{ACC_2006}{accident cases from 2006}
  \item{ACC_2007}{accident cases from 2007}
  \item{ACC_2008}{accident cases from 2008}
  \item{ACC_2009}{accident cases from 2009}
  \item{ACC_2010}{accident cases from 2010}
  \item{ACC_2011}{accident cases from 2011}
  \item{ACC_2012}{accident cases from 2012}
  \item{ACC_2013}{accident cases from 2013}
  \item{ACC_2014}{accident cases from 2014}
  \item{ACC_2015}{accident cases from 2015}
  \item{ACC_2016}{accident cases from 2016}
  \item{ACC_2017}{accident cases from 2017}
  \item{ACC_2018}{accident cases from 2018}
  \item{ACC_2019}{accident cases from 2019}
  \item{ACC_2020}{accident cases from 2020}
  \item{ACC_2021}{accident cases from 2021}
  \item{ACC_2022}{accident cases from 2022}
  \item{ACC_2023}{accident cases from 2023}
  \item{ACC_2024}{accident cases from 2024}
  \item{ACC_2025}{accident cases from 2025}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
accidents
}
\description{
Municipal dataset containing accident cases in Mexico from 2003 to 2025
}
\details{
object sf of a municipal dataset of accident cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
