% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardiovascular_diseases}
\alias{cardiovascular_diseases}
\title{Cardiovascular diseases dataset}
\format{
A sf object with 2477 rows and 26 variables:
\describe{
  \item{CVEGEO}{the ids.}
  \item{NOMBRE}{is the name of the Municipality}
  \item{CD_2003}{cardiovascular cases from 2003}
  \item{CD_2004}{cardiovascular cases from 2004}
  \item{CD_2005}{cardiovascular cases from 2005}
  \item{CD_2006}{cardiovascular cases from 2006}
  \item{CD_2007}{cardiovascular cases from 2007}
  \item{CD_2008}{cardiovascular cases from 2008}
  \item{CD_2009}{cardiovascular cases from 2009}
  \item{CD_2010}{cardiovascular cases from 2010}
  \item{CD_2011}{cardiovascular cases from 2011}
  \item{CD_2012}{cardiovascular cases from 2012}
  \item{CD_2013}{cardiovascular cases from 2013}
  \item{CD_2014}{cardiovascular cases from 2014}
  \item{CD_2015}{cardiovascular cases from 2015}
  \item{CD_2016}{cardiovascular cases from 2016}
  \item{CD_2017}{cardiovascular cases from 2017}
  \item{CD_2018}{cardiovascular cases from 2018}
  \item{CD_2019}{cardiovascular cases from 2019}
  \item{CD_2020}{cardiovascular cases from 2020}
  \item{CD_2021}{cardiovascular cases from 2021}
  \item{CD_2022}{cardiovascular cases from 2022}
  \item{CD_2023}{cardiovascular cases from 2023}
  \item{CD_2024}{cardiovascular cases from 2024}
  \item{CD_2025}{cardiovascular cases from 2025}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
cardiovascular_diseases
}
\description{
Municipal dataset containing cardiovascular cases in Mexico from 2003 to 2025
}
\details{
object sf of a municipal dataset of cardiovascular cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
