% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes dataset}
\format{
A sf object with 2477 rows and 26 variables:
\describe{
  \item{CVEGEO}{the ids.}
  \item{NOMBRE}{is the name of the Municipality}
  \item{DIA_2003}{diabetes cases from 2003}
  \item{DIA_2004}{diabetes cases from 2004}
  \item{DIA_2005}{diabetes cases from 2005}
  \item{DIA_2006}{diabetes cases from 2006}
  \item{DIA_2007}{diabetes cases from 2007}
  \item{DIA_2008}{diabetes cases from 2008}
  \item{DIA_2009}{diabetes cases from 2009}
  \item{DIA_2010}{diabetes cases from 2010}
  \item{DIA_2011}{diabetes cases from 2011}
  \item{DIA_2012}{diabetes cases from 2012}
  \item{DIA_2013}{diabetes cases from 2013}
  \item{DIA_2014}{diabetes cases from 2014}
  \item{DIA_2015}{diabetes cases from 2015}
  \item{DIA_2016}{diabetes cases from 2016}
  \item{DIA_2017}{diabetes cases from 2017}
  \item{DIA_2018}{diabetescases from 2018}
  \item{DIA_2019}{diabetes cases from 2019}
  \item{DIA_2020}{diabetes cases from 2020}
  \item{DIA_2021}{diabetes cases from 2021}
  \item{DIA_2022}{diabetes cases from 2022}
  \item{DIA_2023}{diabetes cases from 2023}
  \item{DIA_2024}{diabetes cases from 2024}
  \item{DIA_2025}{diabetes cases from 2025}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
diabetes
}
\description{
Municipal dataset containing diabetes cases in Mexico from 2003 to 2025
}
\details{
object sf of a municipal dataset of diabetes cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
