% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epoc}
\alias{epoc}
\title{EPOC dataset}
\format{
A sf object with 2477 rows and 26 variables:
\describe{
  \item{CVEGEO}{the ids.}
  \item{NOMBRE}{is the name of the Municipality}
  \item{EPOC_2003}{epoc cases from 2003}
  \item{EPOC_2004}{epoc cases from 2004}
  \item{EPOC_2005}{epoc cases from 2005}
  \item{EPOC_2006}{epoc cases from 2006}
  \item{EPOC_2007}{epoc cases from 2007}
  \item{EPOC_2008}{epoc cases from 2008}
  \item{EPOC_2009}{epoc cases from 2009}
  \item{EPOC_2010}{epoc cases from 2010}
  \item{EPOC_2011}{epoc cases from 2011}
  \item{EPOC_2012}{epoc cases from 2012}
  \item{EPOC_2013}{epoc cases from 2013}
  \item{EPOC_2014}{epoc cases from 2014}
  \item{EPOC_2015}{epoc cases from 2015}
  \item{EPOC_2016}{epoc cases from 2016}
  \item{EPOC_2017}{epoc cases from 2017}
  \item{EPOC_2018}{epoc cases from 2018}
  \item{EPOC_2019}{epoc cases from 2019}
  \item{EPOC_2020}{epoc cases from 2020}
  \item{EPOC_2021}{epoc cases from 2021}
  \item{EPOC_2022}{epoc cases from 2022}
  \item{EPOC_2023}{epoc cases from 2023}
  \item{EPOC_2024}{epoc cases from 2024}
  \item{EPOC_2025}{epoc cases from 2025}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
epoc
}
\description{
Municipal dataset containing EPOC cases in Mexico from 2003 to 2025
}
\details{
object sf of a municipal dataset of epoc cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
