% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hearth_diseases}
\alias{hearth_diseases}
\title{Hearth Diseases dataset}
\format{
A data.frame object with 2478 rows and 24 variables:
\describe{
  \item{CVEGEO}{the ids the municipalities}
  \item{HD_2003}{hearth diseases cases from 2003}
  \item{HD_2004}{hearth diseases cases from 2004}
  \item{HD_2005}{hearth diseases cases from 2005}
  \item{HD_2006}{hearth diseases cases from 2006}
  \item{HD_2007}{hearth diseases cases from 2007}
  \item{HD_2008}{hearth diseases cases from 2008}
  \item{HD_2009}{hearth diseases cases from 2009}
  \item{HD_2010}{hearth diseases cases from 2010}
  \item{HD_2011}{hearth diseases cases from 2011}
  \item{HD_2012}{hearth diseases cases from 2012}
  \item{HD_2013}{hearth diseases cases from 2013}
  \item{HD_2014}{hearth diseases cases from 2014}
  \item{HD_2015}{hearth diseases cases from 2015}
  \item{HD_2016}{hearth diseases cases from 2016}
  \item{HD_2017}{hearth diseases cases from 2017}
  \item{HD_2018}{hearth diseases cases from 2018}
  \item{HD_2019}{hearth diseases cases from 2019}
  \item{HD_2020}{hearth diseases cases from 2020}
  \item{HD_2021}{hearth diseases cases from 2021}
  \item{HD_2022}{hearth diseases cases from 2022}
  \item{HD_2023}{hearth diseases cases from 2023}
  \item{HD_2024}{hearth diseases cases from 2024}
  \item{HD_2025}{hearth diseases cases from 2025}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
hearth_diseases
}
\description{
Municipal dataset containing Hearth Diseases cases in Mexico from 2003 to 2025
}
\details{
object data.frame of a municipal dataset of hearth diseases cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
