% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{influenza_pneumonia}
\alias{influenza_pneumonia}
\title{Influenza-Pneumonia dataset}
\format{
A sf object with 2477 rows and 26 variables:
\describe{
  \item{CVEGEO}{the ids.}
  \item{NOMBRE}{is the name of the Municipality}
  \item{IP_2003}{influenza-pneumonia cases from 2003}
  \item{IP_2004}{influenza-pneumonia cases from 2004}
  \item{IP_2005}{influenza-pneumonia cases from 2005}
  \item{IP_2006}{influenza-pneumonia cases from 2006}
  \item{IP_2007}{influenza-pneumonia cases from 2007}
  \item{IP_2008}{influenza-pneumonia cases from 2008}
  \item{IP_2009}{influenza-pneumonia cases from 2009}
  \item{IP_2010}{influenza-pneumonia cases from 2010}
  \item{IP_2011}{influenza-pneumonia cases from 2011}
  \item{IP_2012}{influenza-pneumonia cases from 2012}
  \item{IP_2013}{influenza-pneumonia cases from 2013}
  \item{IP_2014}{influenza-pneumonia cases from 2014}
  \item{IP_2015}{influenza-pneumonia cases from 2015}
  \item{IP_2016}{influenza-pneumonia cases from 2016}
  \item{IP_2017}{influenza-pneumonia cases from 2017}
  \item{IP_2018}{influenza-pneumonia cases from 2018}
  \item{IP_2019}{influenza-pneumonia cases from 2019}
  \item{IP_2020}{influenza-pneumonia cases from 2020}
  \item{IP_2021}{influenza-pneumonia cases from 2021}
  \item{IP_2022}{influenza-pneumonia cases from 2022}
  \item{IP_2023}{influenza-pneumonia cases from 2023}
  \item{IP_2024}{influenza-pneumonia cases from 2024}
  \item{IP_2025}{influenza-pneumonia cases from 2025}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
influenza_pneumonia
}
\description{
Municipal dataset containing Influenza-Pneumonia cases in Mexico from 2003 to 2025
}
\details{
object sf of a municipal dataset of Influenza-Pneumonia diseases cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
