% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidney_diseases}
\alias{kidney_diseases}
\title{Kidney diseases dataset}
\format{
A data.frame object with 2478 rows and 24 variables:
\describe{
  \item{CVEGEO}{the ids of the municipalities}
  \item{KD_2020}{kidney diseases cases from 2020}
  \item{KD_2021}{kidney diseases cases from 2021}
  \item{KD_2022}{kidney diseases cases from 2022}
  \item{KD_2023}{kidney diseases cases from 2023}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
kidney_diseases
}
\description{
Municipal dataset containing kidney diseases cases in Mexico from 2020 to 2023
}
\details{
object data.frame of a municipal dataset of kidney diseases cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
