% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver_diseases}
\alias{liver_diseases}
\title{Liver diseases dataset}
\format{
A object with 2478 rows and 24 variables:
\describe{
  \item{CVEGEO}{the ids of the municipalities}
  \item{LD_2003}{liver diseases cases from 2003}
  \item{LD_2004}{liver diseases cases from 2004}
  \item{LD_2005}{liver diseases cases from 2005}
  \item{LD_2006}{liver diseases cases from 2006}
  \item{LD_2007}{liver diseases cases from 2007}
  \item{LD_2008}{liver diseases cases from 2008}
  \item{LD_2009}{liver diseases cases from 2009}
  \item{LD_2010}{liver diseases cases from 2010}
  \item{LD_2011}{liver diseases cases from 2011}
  \item{LD_2012}{liver diseases cases from 2012}
  \item{LD_2013}{liver diseases cases from 2013}
  \item{LD_2014}{liver diseases cases from 2014}
  \item{LD_2015}{liver diseases cases from 2015}
  \item{LD_2016}{liver diseases cases from 2016}
  \item{LD_2017}{liver diseases cases from 2017}
  \item{LD_2018}{liver diseases cases from 2018}
  \item{LD_2019}{liver diseases cases from 2019}
  \item{LD_2020}{liver diseases cases from 2020}
  \item{LD_2021}{liver diseases cases from 2021}
  \item{LD_2022}{liver diseases cases from 2022}
  \item{LD_2023}{liver diseases cases from 2023}
  \item{LD_2024}{liver diseases cases from 2024}
  \item{LD_2025}{liver diseases cases from 2025}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
liver_diseases
}
\description{
Municipal dataset containing liver diseases cases in Mexico from 2003 to 2025
}
\details{
object data.frame of a municipal dataset of liver diseases cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
