% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malignant_tumors}
\alias{malignant_tumors}
\title{Malignant tumors  dataset}
\format{
A data.frame object with 2478 rows and 24 variables:
\describe{
  \item{CVEGEO}{the ids of the municipalities}
  \item{MT_2003}{malignant tumors cases from 2003}
  \item{MT_2004}{malignant tumors cases from 2004}
  \item{MT_2005}{malignant tumors cases from 2005}
  \item{MT_2006}{malignant tumors cases from 2006}
  \item{MT_2007}{malignant tumors cases from 2007}
  \item{MT_2008}{malignant tumors cases from 2008}
  \item{MT_2009}{malignant tumors cases from 2009}
  \item{MT_2010}{malignant tumors cases from 2010}
  \item{MT_2011}{malignant tumors cases from 2011}
  \item{MT_2012}{malignant tumors cases from 2012}
  \item{MT_2013}{malignant tumors cases from 2013}
  \item{MT_2014}{malignant tumors cases from 2014}
  \item{MT_2015}{malignant tumors cases from 2015}
  \item{MT_2016}{malignant tumors cases from 2016}
  \item{MT_2017}{malignant tumors cases from 2017}
  \item{MT_2018}{malignant tumors cases from 2018}
  \item{MT_2019}{malignant tumors cases from 2019}
  \item{MT_2020}{malignant tumors cases from 2020}
  \item{MT_2021}{malignant tumors cases from 2021}
  \item{MT_2022}{malignant tumors cases from 2022}
  \item{MT_2023}{malignant tumors cases from 2023}
  \item{MT_2024}{malignant tumors cases from 2024}
  \item{MT_2025}{malignant tumors cases from 2025}
  ...
}
}
\source{
\url{https://pautas.ine.mx/transparencia/mapas/}
}
\usage{
malignant_tumors
}
\description{
Municipal dataset containing malignant tumors cases in Mexico from 2003 to 2025
}
\details{
object data.frame of a municipal dataset of malignant tumors cases in Mexico from 2003 to 2025.
}
\keyword{datasets}
